feat: show icon if has already voted
