+++ packages/shared/lib/core/governance/interfaces/proposal.interface.ts
@@ -3,7 +3,6 @@ import { ProposalStatus } from '../enums'
 import { IOrganization } from './organization.interface'
 
 export interface IProposal {
-    hasVoted?: boolean
     id: string
     milestones?: Record<ProposalStatus, number>
     organization?: IOrganization
