+++ packages/shared/lib/core/governance/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './createProposalsFromEvents'
+export * from './isVotingForProposal'
 export * from './isVotingForSelectedProposal'
